statementAIPArser/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.61.1",
    "pdf2image>=1.17.0",
    "pillow>=11.1.0",
]
